+++ client/src/components/utilities/Toast.tsx
@@ -6,7 +6,6 @@ type Props = {
 };
 
 const Wrapper = styled.div<{ isFading: boolean }>`
-  /* transition: transform 0.6s ease-in-out; */
   animation: toast-in-right 0.6s;
   background: var(--primaryBlue);
   transition: 0.3s ease;
@@ -24,17 +23,7 @@ const Wrapper = styled.div<{ isFading: boolean }>`
     props.isFading &&
     css`
       opacity: 0;
-      /* transform: opacity 2s; */
     `}
-
-  @keyframes toast-in-right {
-    from {
-      transform: translateX(100%);
-    }
-    to {
-      transform: translateX(0);
-    }
-  }
 
   @media screen and (max-width: 375px) {
     font-size: 0.9rem;
@@ -55,7 +44,6 @@ const Toast = ({ message }: Props) => {
     return () => {
       mounted = false;
     };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return <Wrapper isFading={isFading}>{text}</Wrapper>;
